Fix for message headers into delivery reports (#37)
