+++ c_src/erlkaf_consumer.cc
@@ -7,6 +7,7 @@
 #include "erlkaf_logger.h"
 #include "queuemanager.h"
 #include "queuecallbacksdispatcher.h"
+#include "erlkaf_message.h"
 
 #include <vector>
 #include <string>
@@ -180,29 +181,6 @@ rd_kafka_topic_partition_list_t* topic_subscribe(ErlNifEnv* env, ERL_NIF_TERM li
     }
 
     return topics.release();
-}
-
-ERL_NIF_TERM get_headers(ErlNifEnv* env, const rd_kafka_message_t* msg)
-{
-    rd_kafka_headers_t* headers = NULL;
-
-    if(rd_kafka_message_headers(msg, &headers) == RD_KAFKA_RESP_ERR_NO_ERROR && headers)
-    {
-        std::vector<ERL_NIF_TERM> array;
-        array.reserve(rd_kafka_header_cnt(headers));
-
-        const char *name;
-        const void *valuep;
-        size_t size;
-        size_t idx = 0;
-
-        while (!rd_kafka_header_get_all(headers, idx++, &name, &valuep, &size))
-            array.push_back(enif_make_tuple2(env, make_binary(env, name, strlen(name)), make_binary(env, reinterpret_cast<const char*>(valuep), size)));
-
-        return enif_make_list_from_array(env, array.data(), array.size());
-    }
-
-    return ATOMS.atomUndefined;
 }
 
 }  // namespace
@@ -396,14 +374,8 @@ ERL_NIF_TERM enif_consumer_queue_poll(ErlNifEnv* env, int argc, const ERL_NIF_TE
                 first = false;
             }
 
-            ERL_NIF_TERM key = msg->key == NULL ? ATOMS.atomUndefined : make_binary(env, reinterpret_cast<const char*>(msg->key), msg->key_len);
-            ERL_NIF_TERM offset = enif_make_int64(env, msg->offset);
-            ERL_NIF_TERM value = make_binary(env, reinterpret_cast<const char*>(msg->payload), msg->len);
-            ERL_NIF_TERM headers = get_headers(env, msg);
-            ERL_NIF_TERM msg_term = enif_make_tuple7(env, ATOMS.atomMessage, topic, partition, offset, key, value, headers);
-
+            messages.push_back(make_message_nif(env, topic, partition, msg));
             last_offset = msg->offset;
-            messages.push_back(msg_term);
         }
 
         rd_kafka_event_destroy(event);
